app/requirements.txt
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
biopython==1.81
    # via -r requirements.in
black==22.3.0
    # via -r requirements.in
celery[redis]==5.2.7
    # via
    #   -r requirements.in
    #   flower
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coverage[toml]==7.2.3
    # via pytest-cov
django==3.2.18
    # via -r requirements.in
environs==4.1.3
    # via -r requirements.in
exceptiongroup==1.1.1
    # via pytest
faker==17.5.0
    # via -r requirements.in
flake8==4.0.1
    # via
    #   -r requirements.in
    #   flake8-debugger
    #   flake8-isort
    #   flake8-print
    #   pep8-naming
flake8-debugger==4.1.2
    # via -r requirements.in
flake8-isort==4.2.0
    # via -r requirements.in
flake8-print==5.0.0
    # via -r requirements.in
flake8-todo==0.7
    # via -r requirements.in
flower==1.2.0
    # via -r requirements.in
humanize==4.6.0
    # via flower
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
isort==4.3.21
    # via
    #   -r requirements.in
    #   flake8-isort
kombu==5.2.4
    # via celery
levenshtein==0.21.0
    # via -r requirements.in
marshmallow==3.19.0
    # via environs
mccabe==0.6.1
    # via flake8
mock==5.0.2
    # via -r requirements.in
mypy-extensions==1.0.0
    # via black
numpy==1.24.2
    # via biopython
packaging==23.1
    # via
    #   marshmallow
    #   pytest
pathspec==0.11.1
    # via black
pep8-naming==0.13.1
    # via -r requirements.in
platformdirs==3.2.0
    # via black
pluggy==1.0.0
    # via pytest
prometheus-client==0.16.0
    # via flower
prompt-toolkit==3.0.38
    # via click-repl
psycopg2==2.8.5
    # via -r requirements.in
pycodestyle==2.8.0
    # via
    #   flake8
    #   flake8-debugger
    #   flake8-print
    #   flake8-todo
pydocstyle==3.0.0
    # via -r requirements.in
pyflakes==2.4.0
    # via flake8
pytest==7.3.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-mock
pytest-cov==4.0.0
    # via -r requirements.in
pytest-mock==3.10.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
python-dotenv==1.0.0
    # via environs
pytz==2023.3
    # via
    #   celery
    #   django
    #   flower
rapidfuzz==3.0.0
    # via levenshtein
redis==4.5.4
    # via celery
requests==2.28.2
    # via -r requirements.in
retry-decorator==1.1.1
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   pydocstyle
    #   python-dateutil
snowballstemmer==2.2.0
    # via pydocstyle
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   pytest
tornado==6.3
    # via flower
urllib3==1.26.15
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
